requirements.txt
numpy==1.24.4
pandas==2.0.3
PyYAML==5.3.1
scikit-learn==1.3.2
scipy==1.10.1
timm==0.9.16
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
